Fix utterances color scheme setting in the preferred mode
Reviously we set utterances' theme to `preferred-color-scheme`
under `eureka` theme. It means using the browser default color
scheme. It's fine if the default was light mode as we have
logic to switch dark mode. However, if the deafult was the dark
mode, utterances won't change its theme even if light mode is
explicity turned on for the website. This commit fixes it.